Fixed circular import issue
